Start erroring out for unsupported lateral subqueries
With the introduction of #4385 we inadvertently started allowing and
pushing down certain lateral subqueries that were unsafe to push down.
To be precise the type of LATERAL subqueries that is unsafe to push down
has all of the following properties:
1. The subquery references a recurring tuple (e.g. a reference table) from
   outside of the subquery
2. The subquery requires a merge step (e.g. a LIMIT)
3. The reference to the recurring tuple should be something else than an
   equality check on the distribution column, e.g. equality on a non
   distribution column.

Property number three is considered both hard to detect and probably not
used very often. Thus this PR ignores property number three and causes
query planning to error out if the first two properties hold.

Fixes #5327