Use gpids in deadlock detector
The deadlock detector had a bug where it would detect deadlocks even
when there were none. This could happen when the lock graph contains a
combination of transactions had a distributed transaction id assigned,
but also some transactions that did not.

Multiple different transactions without a transaction id were considered
the same transaction (with transaction id 0). Thus a cycle would be
instantly detected if two of those transactions were waiting on
eachother. However, due to the nature of the deadlock detector these
transaction without distributed transaction ids would normally not be
included in the lock graph at all. Only when a transaction that has a
distributed transaction id was waiting on one without, would it be
included.

Fixes #1803