Fix flaky test in a better way
The reason the test was only sometimes failing is that previous DDL
commands triggered propagation of the `search_path` value on the cached
connections. By disabling connection caching the tests started failing
consistently. This changed reverts the previous fix in #5894, and
instead disables connection caching and updates the expected output
accordingly.

This is related to #462, because session level `search_path` is not
propagated to workers. It's actually possible to create real failures,
by not schema-qualifying the tables in the function body.